refactor: rename "CacheService" directory.
There is a "CacheService.cs" file within the "CacheService" directory, so the directory was renamed and a namespace added to match the folder structure.
