+++ src/LibraryManager/Cache/CacheService.cs
@@ -10,7 +10,7 @@ using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Contracts.Caching;
 using Microsoft.Web.LibraryManager.Utilities;
 
-namespace Microsoft.Web.LibraryManager
+namespace Microsoft.Web.LibraryManager.Cache
 {
     /// <summary>
     /// Service to manage basic operations on libraries cache
