+++ src/libman/EnvironmentSettings.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.IO;
+using Microsoft.Web.LibraryManager.Cache;
 using Microsoft.Web.LibraryManager.Contracts;
 using Microsoft.Web.LibraryManager.Tools.Contracts;
